backend/requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.22.0
sqlalchemy==2.0.21
alembic==1.13.2
celery==5.4.0
redis==5.0.7
python-dotenv==1.0.0
requests==2.32.0
beautifulsoup4==4.12.3
trafilatura==1.12.0
pydantic==2.7.1
pydantic-settings==2.2.1
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
psycopg2-binary==2.9.9
argon2-cffi==23.1.0
google-search-results==2.4.2
dateparser==1.2.0

transformers==4.40.1
torch
numpy

pydantic[email]